refactor: split out more tests from `process_blocks.rs`
