+++ integration-tests/src/tests/client/features.rs
@@ -4,3 +4,4 @@
 mod account_id_in_function_call_permission;
 #[cfg(feature = "protocol_feature_fix_contract_loading_cost")]
 mod fix_contract_loading_cost;
+mod restore_receipts_after_fix;
